requirements.txt
asgiref==3.2.7
certifi==2019.11.28
click==7.1.1
Django==3.0.4
Flask==1.1.1
image==1.5.28
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.0.0
pytesseract==0.3.3
python-dotenv==0.12.0
pytz==2019.3
sqlparse==0.3.1
Werkzeug==1.0.0
wincertstore==0.2
gunicorn==19.9.0
